+++ Pickle/Classes/PhotoGalleryViewController.swift
@@ -81,7 +81,18 @@ internal final class PhotoGalleryViewController: UIViewController,
     internal private(set) lazy var fetchResult: PHFetchResult<PHAsset> = {
         let options = PHFetchOptions()
         options.sortDescriptors = [NSSortDescriptor(key: "creationDate", ascending: false)]
-        options.predicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.image.rawValue)
+
+        switch configuration?.mediaType {
+        case .all?:
+            options.predicate = nil
+        case .image?:
+            options.predicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.image.rawValue)
+        case .video?:
+            options.predicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.video.rawValue)
+        default:
+            options.predicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.image.rawValue)
+        }
+
         return PHAsset.fetchAssets(in: self.album, options: options)
     }()
 
@@ -108,7 +119,8 @@ internal final class PhotoGalleryViewController: UIViewController,
         collectionView.delegate = self
         collectionView.backgroundColor = UIColor.white
         collectionView.register(PhotoGalleryCameraCell.self, forCellWithReuseIdentifier: String(describing: PhotoGalleryCameraCell.self))
-        collectionView.register(PhotoGalleryCell.self, forCellWithReuseIdentifier: String(describing: PhotoGalleryCell.self))
+        collectionView.register(GalleryPhotoCell.self, forCellWithReuseIdentifier: String(describing: GalleryPhotoCell.self))
+        collectionView.register(GalleryVideoCell.self, forCellWithReuseIdentifier: String(describing: GalleryVideoCell.self))
         collectionView.register(PhotoGalleryLiveViewCell.self, forCellWithReuseIdentifier: String(describing: PhotoGalleryLiveViewCell.self))
         collectionView.allowsMultipleSelection = true
         return collectionView
@@ -164,15 +176,22 @@ internal final class PhotoGalleryViewController: UIViewController,
             }
         }
 
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: String(describing: PhotoGalleryCell.self), for: indexPath)
         let index = isCameraCompatible ? indexPath.row - 1 : indexPath.row
         let asset = fetchResult[index]
 
-        if let text = delegate?.photoGalleryViewController(self, taggedTextForPhoto: asset) {
-            (cell as? PhotoGalleryCell)?.configure(with: asset, taggedText: text, configuration: configuration)
+        let text = delegate?.photoGalleryViewController(self, taggedTextForPhoto: asset)
+        var cell: UICollectionViewCell
+        if asset.mediaType == .video {
+            cell = collectionView.dequeueReusableCell(withReuseIdentifier: String(describing: GalleryVideoCell.self), for: indexPath)
+            (cell as? GalleryVideoCell)?.configure(with: asset, taggedText: text, configuration: configuration)
+        } else {
+            cell = collectionView.dequeueReusableCell(withReuseIdentifier: String(describing: GalleryPhotoCell.self), for: indexPath)
+            (cell as? GalleryPhotoCell)?.configure(with: asset, taggedText: text, configuration: configuration)
+        }
+
+        if text != nil {
             collectionView.selectItem(at: indexPath, animated: false, scrollPosition: [])
         } else {
-            (cell as? PhotoGalleryCell)?.configure(with: asset, configuration: configuration)
             collectionView.deselectItem(at: indexPath, animated: false)
         }